backend/requirements.txt
# ── Runtime ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0        # required by FastAPI TestClient (starlette)
